Add files via upload
I used Machine learning techniques, train and test datasets of the air pollution and I predicted the values of the pollutants of the test dataset using scikit-learn library. I made a http server in python and sent the results of prediction to the android application using the http server. The android application connects to the server and shows the values of the pollutants in different dates as a list to the users.